+++ tcl/tc-sc-interaction/SecurityPolicyRulesTest.cc
@@ -2836,15 +2836,15 @@ TEST_P(SecurityPolicyRulesSetProperty, PolicyRulesSC)
     if (GetParam().proxyObjAllowedToCallSetProperty && GetParam().busObjAllowedToRespondToSetPropertyCall) {
         EXPECT_EQ(ER_OK, setPropertyStatus);
         //Verify Prop1 is changed.
-        EXPECT_EQ(static_cast<uint32_t>(513), TCBus.ReadProp1());
+        EXPECT_EQ(513, TCBus.ReadProp1());
     } else if (GetParam().proxyObjAllowedToCallSetProperty && !GetParam().busObjAllowedToRespondToSetPropertyCall) {
         EXPECT_EQ(ER_PERMISSION_DENIED, setPropertyStatus);
         //Currently no way to find out that the error string is org.alljoyn.Bus.Security.Error.PermissionDenied
         //Verify Prop1 is unchanged (note prop1 defaults to 42 by the constructor)
-        EXPECT_EQ(static_cast<uint32_t>(42), TCBus.ReadProp1());
+        EXPECT_EQ(42, TCBus.ReadProp1());
     } else { //!GetParam().proxyObjAllowedToCallSetProperty
         EXPECT_EQ(ER_PERMISSION_DENIED, setPropertyStatus);
-        EXPECT_EQ(static_cast<uint32_t>(42), TCBus.ReadProp1());
+        EXPECT_EQ(42, TCBus.ReadProp1());
     }
 
     /* clean up */
@@ -3164,10 +3164,10 @@ TEST_P(SecurityPolicyRulesSetPropertyManifest, PolicyRulesSC)
     if (GetParam().proxyObjAllowedToCallSetProperty && GetParam().busObjAllowedToRespondToSetPropertyCall) {
         EXPECT_EQ(ER_OK, setPropertyStatus);
         //Verify Prop1 is changed.
-        EXPECT_EQ(static_cast<uint32_t>(513), TCBus.ReadProp1());
+        EXPECT_EQ(513, TCBus.ReadProp1());
     } else { //!GetParam().proxyObjAllowedToCallSetProperty
         EXPECT_TRUE(ER_PERMISSION_DENIED == setPropertyStatus || ER_BUS_REPLY_IS_ERROR_MESSAGE == setPropertyStatus);
-        EXPECT_EQ(static_cast<uint32_t>(42), TCBus.ReadProp1());
+        EXPECT_EQ(42, TCBus.ReadProp1());
     }
 
     /* clean up */
@@ -6282,7 +6282,7 @@ TEST_F(SecurityPolicyRulesTest, PolicyRules_DENY_1_SC)
     MsgArg prop1Arg;
     EXPECT_EQ(ER_OK, prop1Arg.Set("i", 513));
     EXPECT_EQ(ER_OK, proxy.SetProperty(interfaceName, "Prop1", prop1Arg));
-    EXPECT_EQ(static_cast<uint32_t>(513), TCBus.ReadProp1());
+    EXPECT_EQ(513, TCBus.ReadProp1());
 
     MsgArg prop1ArgOut;
     EXPECT_EQ(ER_OK, proxy.GetProperty(interfaceName, "Prop1", prop1Arg));
@@ -6607,7 +6607,7 @@ TEST_F(SecurityPolicyRulesTest, PolicyRules_DENY_2_SC)
     MsgArg prop1Arg;
     EXPECT_EQ(ER_OK, prop1Arg.Set("i", 513));
     EXPECT_EQ(ER_OK, proxy.SetProperty(interfaceName, "Prop1", prop1Arg));
-    EXPECT_EQ(static_cast<uint32_t>(513), TCBus.ReadProp1());
+    EXPECT_EQ(513, TCBus.ReadProp1());
 
     MsgArg prop1ArgOut;
     EXPECT_EQ(ER_OK, proxy.GetProperty(interfaceName, "Prop1", prop1Arg));
@@ -6925,7 +6925,7 @@ TEST_F(SecurityPolicyRulesTest, PolicyRules_DENY_3_SC)
     MsgArg prop1Arg;
     EXPECT_EQ(ER_OK, prop1Arg.Set("i", 513));
     EXPECT_EQ(ER_OK, proxy.SetProperty(interfaceName, "Prop1", prop1Arg));
-    EXPECT_EQ(static_cast<uint32_t>(513), TCBus.ReadProp1());
+    EXPECT_EQ(513, TCBus.ReadProp1());
 
     MsgArg prop1ArgOut;
     EXPECT_EQ(ER_OK, proxy.GetProperty(interfaceName, "Prop1", prop1Arg));
@@ -7241,7 +7241,7 @@ TEST_F(SecurityPolicyRulesTest, PolicyRules_DENY_4_SC)
     MsgArg prop1Arg;
     EXPECT_EQ(ER_OK, prop1Arg.Set("i", 513));
     EXPECT_EQ(ER_OK, proxy.SetProperty(interfaceName, "Prop1", prop1Arg));
-    EXPECT_EQ(static_cast<uint32_t>(513), TCBus.ReadProp1());
+    EXPECT_EQ(513, TCBus.ReadProp1());
 
     MsgArg prop1ArgOut;
     EXPECT_EQ(ER_OK, proxy.GetProperty(interfaceName, "Prop1", prop1Arg));
@@ -8478,7 +8478,7 @@ TEST_F(SecurityPolicyRulesTest, PolicyRules_DENY_7_SC)
     MsgArg prop1Arg;
     EXPECT_EQ(ER_OK, prop1Arg.Set("i", 513));
     EXPECT_EQ(ER_OK, proxy.SetProperty(interfaceName, "Prop1", prop1Arg));
-    EXPECT_EQ(static_cast<uint32_t>(513), TCBus.ReadProp1());
+    EXPECT_EQ(513, TCBus.ReadProp1());
 
     MsgArg prop1ArgOut;
     EXPECT_EQ(ER_OK, proxy.GetProperty(interfaceName, "Prop1", prop1Arg));
@@ -8805,7 +8805,7 @@ TEST_F(SecurityPolicyRulesTest, PolicyRules_DENY_8_SC)
     MsgArg prop1Arg;
     EXPECT_EQ(ER_OK, prop1Arg.Set("i", 513));
     EXPECT_EQ(ER_OK, proxy.SetProperty(interfaceName, "Prop1", prop1Arg));
-    EXPECT_EQ(static_cast<uint32_t>(513), TCBus.ReadProp1());
+    EXPECT_EQ(513, TCBus.ReadProp1());
 
     MsgArg prop1ArgOut;
     EXPECT_EQ(ER_OK, proxy.GetProperty(interfaceName, "Prop1", prop1Arg));
@@ -9815,7 +9815,7 @@ TEST_F(SecurityPolicyRulesTest, PolicyRules_DENY_11_SC)
     MsgArg prop1Arg;
     EXPECT_EQ(ER_OK, prop1Arg.Set("i", 513));
     EXPECT_EQ(ER_OK, proxy.SetProperty(interfaceName, "Prop1", prop1Arg));
-    EXPECT_EQ(static_cast<uint32_t>(513), TCBus.ReadProp1());
+    EXPECT_EQ(513, TCBus.ReadProp1());
 
     MsgArg prop1ArgOut;
     EXPECT_EQ(ER_OK, proxy.GetProperty(interfaceName, "Prop1", prop1Arg));
@@ -10289,7 +10289,7 @@ TEST_F(SecurityPolicyRulesTest, PolicyRules_DENY_12_SC)
     MsgArg prop1Arg;
     EXPECT_EQ(ER_OK, prop1Arg.Set("i", 513));
     EXPECT_EQ(ER_OK, proxy.SetProperty(interfaceName, "Prop1", prop1Arg));
-    EXPECT_EQ(static_cast<uint32_t>(513), TCBus.ReadProp1());
+    EXPECT_EQ(513, TCBus.ReadProp1());
 
     MsgArg prop1ArgOut;
     EXPECT_EQ(ER_OK, proxy.GetProperty(interfaceName, "Prop1", prop1Arg));
@@ -10718,7 +10718,7 @@ TEST_F(SecurityPolicyRulesTest, PolicyRules_DENY_13_SC)
     MsgArg prop1Arg;
     EXPECT_EQ(ER_OK, prop1Arg.Set("i", 513));
     EXPECT_EQ(ER_OK, proxy.SetProperty(interfaceName, "Prop1", prop1Arg));
-    EXPECT_EQ(static_cast<uint32_t>(513), TCBus.ReadProp1());
+    EXPECT_EQ(513, TCBus.ReadProp1());
 
     MsgArg prop1ArgOut;
     EXPECT_EQ(ER_OK, proxy.GetProperty(interfaceName, "Prop1", prop1Arg));
@@ -11098,7 +11098,7 @@ TEST_F(SecurityPolicyRulesTest, PolicyRules_DENY_14_SC)
     MsgArg prop1Arg;
     EXPECT_EQ(ER_OK, prop1Arg.Set("i", 513));
     EXPECT_EQ(ER_OK, proxy.SetProperty(interfaceName, "Prop1", prop1Arg));
-    EXPECT_EQ(static_cast<uint32_t>(513), TCBus.ReadProp1());
+    EXPECT_EQ(513, TCBus.ReadProp1());
 
     MsgArg prop1ArgOut;
     EXPECT_EQ(ER_OK, proxy.GetProperty(interfaceName, "Prop1", prop1Arg));
@@ -11408,7 +11408,7 @@ TEST_F(SecurityPolicyRulesTest, PolicyRules_DENY_15_SC)
     MsgArg prop1Arg;
     EXPECT_EQ(ER_OK, prop1Arg.Set("i", 513));
     EXPECT_EQ(ER_OK, proxy.SetProperty(interfaceName, "Prop1", prop1Arg));
-    EXPECT_EQ(static_cast<uint32_t>(513), TCBus.ReadProp1());
+    EXPECT_EQ(513, TCBus.ReadProp1());
 
     MsgArg prop1ArgOut;
     EXPECT_EQ(ER_OK, proxy.GetProperty(interfaceName, "Prop1", prop1Arg));
@@ -11638,7 +11638,7 @@ TEST_F(SecurityPolicyRulesTest, PolicyRules_DENY_16_SC)
     MsgArg prop1Arg;
     EXPECT_EQ(ER_OK, prop1Arg.Set("i", 513));
     EXPECT_EQ(ER_OK, proxy.SetProperty(interfaceName, "Prop1", prop1Arg));
-    EXPECT_EQ(static_cast<uint32_t>(513), TCBus.ReadProp1());
+    EXPECT_EQ(513, TCBus.ReadProp1());
 
     MsgArg prop1ArgOut;
     EXPECT_EQ(ER_OK, proxy.GetProperty(interfaceName, "Prop1", prop1Arg));
